requirements.txt
Pillow
numpy
scipy
opencv-python
opencv-contrib-python
higra
matplotlib
utils

